guava optional

+++ lens-cube/src/test/java/org/apache/lens/cube/metadata/TestCubeMetastoreClient.java
@@ -439,7 +439,7 @@ public class TestCubeMetastoreClient {
     CubeDimAttribute newzipdim = altered.getAttributeByName("newZipDim");
     Assert.assertTrue(newzipdim instanceof BaseDimAttribute);
     Assert.assertEquals(((BaseDimAttribute) newzipdim).getType(), "int");
-    Assert.assertEquals((((BaseDimAttribute) newzipdim).getNumOfDistinctValues()).longValue(), (long)1000);
+    Assert.assertEquals((((BaseDimAttribute) newzipdim).getNumOfDistinctValues()), 1000L);
 
     CubeDimAttribute newrefdim = altered.getAttributeByName("newRefDim");
     Assert.assertTrue(newrefdim instanceof ReferencedDimAtrribute);
